+++ libgimp/gimppath_pdb.h
@@ -38,7 +38,7 @@ GimpPath*          gimp_path_new_from_text_layer       (GimpImage            *im
                                                         GimpLayer            *layer);
 GimpPath*          gimp_path_copy                      (GimpPath             *path);
 gint*              gimp_path_get_strokes               (GimpPath             *path,
-                                                        gint                 *num_strokes);
+                                                        gsize                *num_strokes);
 gdouble            gimp_path_stroke_get_length         (GimpPath             *path,
                                                         gint                  stroke_id,
                                                         gdouble               precision);
@@ -81,18 +81,18 @@ gboolean           gimp_path_stroke_flip_free          (GimpPath             *pa
                                                         gdouble               y2);
 GimpPathStrokeType gimp_path_stroke_get_points         (GimpPath             *path,
                                                         gint                  stroke_id,
-                                                        gint                 *num_points,
+                                                        gsize                *num_points,
                                                         gdouble             **controlpoints,
                                                         gboolean             *closed);
 gint               gimp_path_stroke_new_from_points    (GimpPath             *path,
                                                         GimpPathStrokeType    type,
-                                                        gint                  num_points,
+                                                        gsize                 num_points,
                                                         const gdouble        *controlpoints,
                                                         gboolean              closed);
 gdouble*           gimp_path_stroke_interpolate        (GimpPath             *path,
                                                         gint                  stroke_id,
                                                         gdouble               precision,
-                                                        gint                 *num_coords,
+                                                        gsize                *num_coords,
                                                         gboolean             *closed);
 gint               gimp_path_bezier_stroke_new_moveto  (GimpPath             *path,
                                                         gdouble               x0,
